requirements.txt
ten-vad
tornado
requests
websocket-client
soundfile
numpy
matplotlib
librosa
